KRACOEUS-6956 : repack propdev budget fix upon comments
