+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/budget/auth/task/BudgetProposalTaskFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.coeus.propdev.impl.budget.auth;
+package org.kuali.coeus.propdev.impl.budget.auth.task;
 
 import org.apache.struts.action.ActionForm;
 import org.kuali.coeus.sys.framework.auth.task.Task;
