+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/budget/auth/task/RejectProposalBudgetAuthorizer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.coeus.propdev.impl.budget.action;
+package org.kuali.coeus.propdev.impl.budget.auth.task;
 
 import org.kuali.coeus.sys.framework.service.KcServiceLocator;
 import org.kuali.coeus.sys.framework.workflow.KcDocumentRejectionService;
