--- coeus-code/src/main/java/org/kuali/coeus/propdev/impl/budget/docperm/ViewProposalBudgetPersonSalaryAuthorizer.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2005-2014 The Kuali Foundation
- * 
- * Licensed under the Educational Community License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- * http://www.opensource.org/licenses/ecl1.php
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kuali.coeus.propdev.impl.budget.docperm;
-
-import org.kuali.coeus.propdev.impl.core.ProposalDevelopmentDocument;
-import org.kuali.kra.budget.document.BudgetDocument;
-import org.kuali.kra.budget.document.authorization.BudgetTask;
-import org.kuali.kra.budget.document.authorizer.BudgetAuthorizer;
-import org.kuali.kra.infrastructure.Constants;
-import org.kuali.kra.infrastructure.PermissionConstants;
-
-public class ViewProposalBudgetPersonSalaryAuthorizer extends BudgetAuthorizer{
-
-    public boolean isAuthorized(String userId, BudgetTask task) {
-
-        BudgetDocument budgetDocument = task.getBudgetDocument();
-        ProposalDevelopmentDocument doc = (ProposalDevelopmentDocument) budgetDocument.getParentDocument();
-        
-        return hasParentPermission(userId, doc, Constants.MODULE_NAMESPACE_BUDGET, PermissionConstants.VIEW_SALARIES);
-    }
-}
-
